+++ CHANGELOG.md
@@ -22,11 +22,11 @@ All notable changes to the "Angular Toolbox" extension will be documented in thi
 
 ### Chore
 
-- **extension metadata**: add extension icon and select final name.
+- **extension metadata**: add extension icon and select final name.([91f44f6](https://github.com/Fernandocgomez/vscode-extensions.angular-toolbox/pull/22/commits/91f44f66021aba7b2fcf6650a5b5bf5e16a71588))
 
 ### Changes
 
-- **templates**: change user custom templates and custom config folder name from ".angular-custom-templates" to ".angular-toolbox"
+- **templates**: change user custom templates and custom config folder name from ".angular-custom-templates" to ".angular-toolbox" ([6d46786](https://github.com/Fernandocgomez/vscode-extensions.angular-toolbox/pull/22/commits/6d46786b5b2e424c43b98f447037af3fc51ba606))
 
 ## [Released]